Generate brandable logos/images Qt resource file at cmake time
The window icon and the about image are brandable. We are following
ParaView on this (effectively using a macro from them) and generate the
Qt resource file to incorporate the window icon and the about image at
cmake ("configure") time instead of including them in the main resource
file "screen.qrc".