integrate GetGitRevisionDescription
The behaviour for setting the Sepiola version is now as follows:

if `git` is installed and it is valid git repository:
    VERSION = ${LAST_TAG}[-${NUMBER OF COMMITS SINCE LAST TAG}-${CURRENT_HASH}[-dirty]]

    where the part after the ${LAST_TAG} is empty if the current commit
    is tagged with the version itself, and therefore equal to the
    version hardcoded in CMakeLists.txt

    the `-dirty` flag is added if the current git checkout has
    uncommitted changes

else:
    use the version from CMakeLists.txt

    ... and therefore assume that it is built from a release tarball.

    It is still possible that this is not the case and someone used a
    tarball from the develop branch instead. But this way the release
    process is easy enough for us.

It is important to use the modified git-flow process as
implemented by the git-flow (AVH Edition) from
https://github.com/petervanderdoes/gitflow to ensure that the release
tags end up in all branches (and especially `develop`).

It is still necessary to update the version in `CMakeLists.txt` in the
release process.

Further refinement to the version derivation could be done by evaluating
the branch. In case of a release branch, one could set the version to
X.Y.Z_preN as extracted from branch `release/vX.Y.Z`