installer: do not add version to install directory
This is related to #37. After this change it will not be possible
anymore to have different versions of Sepiola installed at the same
time (different reseller versions still work since they have a different
name). On the other hand most people do not want that and those people
who want that (for example for testing) know how to work around.

It solves the problem of invalid scheduler entries after updating
Sepiola.