installer: explicitly delete pulled-in libraries
The CPack generated list of files to uninstall includes only the files
installed from the $INST_DIR, but we pull in files directly from the
MinGW directory which therefore do not get recorded and do not end up in
the list of files to be uninstalled. We therefore have to explicitly
uninstall them.

As an alternative we could instead try to add those DLLs to the INSTALL command
in src/CMakeLists.txt

This fixes #36.