model/gui: use QFileSystemModel instead of QDirModel
QDirModel is deprecated and the default behaviour of QFileSystemModel
seems to match the current behaviour of our QDirModel setup. Replace it.

The advantage is that QFileSystemModel walks the tree in a separate
thread and should not block the UI. It furthermore has notify support
and therefore no manual refresh is needed anymore.